+++ docs/demos/multi-page-basics/app_dbc.py
@@ -1,10 +1,11 @@
-
 import dash
 import dash_labs as dl
 import dash_bootstrap_components as dbc
 
 
-app = dash.Dash(__name__, plugins=[dl.plugins.pages], external_stylesheets=[dbc.themes.BOOTSTRAP])
+app = dash.Dash(
+    __name__, plugins=[dl.plugins.pages], external_stylesheets=[dbc.themes.BOOTSTRAP]
+)
 
 navbar = dbc.NavbarSimple(
     dbc.DropdownMenu(
@@ -23,7 +24,10 @@ navbar = dbc.NavbarSimple(
 )
 
 app.layout = dbc.Container(
-    [navbar, dl.plugins.page_container, ],
+    [
+        navbar,
+        dl.plugins.page_container,
+    ],
     className="dbc",
     fluid=True,
 )