+++ docs/demos/multi-page-example1/pages/heatmaps.py
@@ -1,5 +1,5 @@
-
 import dash
+
 dash.register_page(__name__, path="/")
 
 from dash import Dash, dcc, html, Input, Output, callback