+++ docs/demos/multi-page-nested-folders/pages/chapter/bar-charts.py
@@ -1,8 +1,7 @@
 import dash
 
-dash.register_page(__name__)
+dash.register_page(__name__, icon="fas fa-chart-bar")
 
-print(__name__)
 
 from dash import Dash, dcc, html, Input, Output, callback
 import plotly.express as px