+++ docs/demos/multi-page-nested-folders/pages/heatmaps.py
@@ -1,7 +1,7 @@
-
 import dash
-dash.register_page(__name__, path="/")
-print(__name__)
+
+dash.register_page(__name__)
+
 
 from dash import Dash, dcc, html, Input, Output, callback
 import plotly.express as px