+++ docs/demos/multi-page-nested-folders/pages/histograms.py
@@ -1,6 +1,6 @@
 import dash
 
-dash.register_page(__name__)
+dash.register_page(__name__, path="/")
 
 from dash import Dash, dcc, html, Input, Output, callback
 import plotly.express as px